GUI and CLI versions of deveny_grangle
Now contains both the GUI and CLI versions of deveny_grangle.  When run from
the command line, the user can select which version of the routine they would
like.  Both utilize the same functional subroutines for the actual computation,
so it is strictly an interface preference for the situation.

The GUI version does do some error checking for entering a proper numerical
central wavelength within the range of the optical capabilities of the CCD
that the CLI version does not.

The CLI version is a direct translation of the extant IDL version, and should
appear no different than what the user expects.

	modified:   deveny_grangle.py