Add calculate wavelength from grating tilt angle
In the GUI version of the calculator, add the ability to compute the central
wavelength from a given grating angle.  To avoid confusion when a user enters
both inputs (lambda + grangle), the routine updates the other input with the
computed value.

The GUI may look a little cluttered, with two separate "Compute" buttons, but
hopefully the utility will be worth it.

	modified:   deveny_grangle.py