Two versions of the GUI available
Previous comits created a GUI that allowed for both the computation of the
grating angle given central wavelenth (original purpose), and the central
wavelength given a grating angle.  This GUI, while useful in some situations,
is visually overbusy.

The original-purpose GUI has been restored, and is both the default (no
command line argumnents) or available via the 'GUI' command line argument.
The visually overbusy GUI is still available, but is invoked using the 'MAX'
command line argument.

The CLI interface is unchanged.

	modified:   deveny_grangle.py