Continued porting in IDL functions for dfocus
There are a lot of individual functions that are used for dfocus!  Continued
adding those as I work down through the main dfocus() routine.

The current python code here is pretty much a direct translation of the IDL
source.  Once it is all in and functioning, I'll work on streamlining it and
making it work in a more pythonic fashion.

The code in this commit runs without error... not the entire dfocus() routine
yet, and the content has not been rigorously checked, but runs without error.
More progress to come.

	modified:   dfocus.py