dfocus() is in and runs without error
The routine dfocus() is conerted into python, and runs without error.

When run, it does not output the same values as the IDL counterpart, so
investigation about the values propagated through the code needs to happen
next.

It also does not do any of the plotting yet.  Once I am satisfied that the
output values match, then I can add the plotting.

	modified:   dfocus.py