Current status
dfocus (python) is not finding all of the lines in a sample data set that the
IDL version found.  (Yay!)  That means the find_lines() routine is up to snuff.

The line fitting and focus-curve fitting routines still leave a little to be
desired.

Wanted to commit this to be able to see what changes I am making  for the
remaining pieces.

	modified:   __init__.py
	modified:   deveny_idl/dfocus/dfitcurves.pro
	modified:   dfocus.py