Add a general-purpose function that wraps key elements
This is what got me to dig into DFOCUS in the first place...  I wanted a helper
function to return the line centers for a given image filename as part of my
flexure analysis.  This is that helper function.  I guess, along the way, I
needed to get DFOCUS actually running and fully functional.  Yay!

The aforementioned testing and GUI stuff still need to happen, but not now.

	modified:   dfocus.py