Add port of Phil Massey's LMI ETC
Ported the ETC from PHP.  At present, the `etc_calc.py` module contains only the
funtions needed to compute exposure times... there is no UI as such.

The goal will be to add some sort of UI (whether graphical or not is to be seen)
to make this a useful tool to have on dct-obs1/dct-obs2.

The base information about the filters (sky counts, etc.) that were hard-coded
in the PHP file are contained in the `.ecsv` file accompanying.

Small changes to the other two `.py` files to clarify the package purpose.

	modified:   deveny_grangle.py
	modified:   dfocus.py
	new file:   etc_calc.py
	new file:   etc_filter_info.ecsv