Add broker listening to `deveny_collfocus`
If run on `dct-obs1` or `dct-obs2`, the `deveny_collfocus` GUI can request the
current mount temperature and grating angle from the ActiveMQ broker at the site
and place those values in the appropriate input boxes.  This is a convenience
feature, but the GUI works fine without broker connection where the user needs
to enter all values by hand.

	modified:   .gitignore
	modified:   README.rst
	new file:   doc/api/obstools.broker_listener.rst
	new file:   doc/api/obstools.deveny_collfocus.rst
	modified:   doc/api/obstools.rst
	new file:   obstools/broker_listener.py
	modified:   obstools/deveny_collfocus.py
	modified:   obstools/utils.py
	modified:   setup.cfg