Add iterative fitting; deal with `objmodel`
For iterative fitting, find rows where the RMS of the fit is
significantly larger than the mean and refit those lines with tighter
bounds and initial guesses based on the fits of adjacent "good" lines.

For the `objmodel`, sometimes the PypeIt-extracted object model contains
sinusoidal signal.  If this is the case, then fitting the pattern to the
residual image is not correct because some of the sinusoid has been
removed by PypeIt.  In these cases, add the object model back into
residual image before fitting line-by-line.

	modified:   obstools/scrub_deveny_pickup.py
	modified:   setup.cfg