Remove `dfocus` ability to move focus frames
This was a feature added to simplify the PypeIt frametyping bit so that
focus frames just didn't appear in the output `.pypeit` file.  The
DRP has since been modified to comment out all frames with `None` as
the frametype (which is the case for DeVeny Focus frames).

So, since moving the focus files is no longer needed, plus there was
some issue reported by one of the BU observers about this, I am simply
removing the move feature.

	modified:   doc/help/dfocus.rst
	modified:   obstools/dfocus.py