LMI ETC calculations match Phil's PHP code
The LMI ETC GUI's caclulations match those from Phil Massy's ETC
webpage.  It is possible to add the results from a calculation into an
AstroPy table for saving (and comparing different setups and requirements).

Next steps will be to show the table in a window and offer to save it
in one or more formats to disk.

	modified:   obstools/ETCWindow.py
	modified:   obstools/ETCWindow.ui
	modified:   obstools/lmi_etc.py