Force unicode read on astropy ECSV table
This fixes a CI error for Windows where the ECSV filter list file was
not being properly read in as unicode (utf-8).  This only matters
because two of the filters are related to the Hα line at 6563Å, and I
stubbornly included the unicode character.  I will note that the macOS
and ubuntu GitHub runners had no problem correctly using unicode.

	modified:   obstools/data/etc_filter_info.ecsv
	modified:   obstools/lmi_etc.py